# openapi.yaml
openapi: 3.0.3
info:
  title: Chatbot AI Vision API
  description: |
    API for AI-powered chatbot with image analysis capabilities using Google Gemini.
    
    ## Features
    - Real-time chat with AI using streaming responses
    - Image upload and analysis
    - Multi-turn conversations with context
    - Advanced AI configuration options
    
    ## Base URLs
    - **Production**: `https://chat.fsu.my.id/api`
    - **Development**: `http://localhost:3000/api`
    
    ## Rate Limiting
    All endpoints are rate-limited per IP address:
    - Chat: 20 requests/minute
    - Image Upload: 10 requests/minute
    - Image Analysis: 15 requests/minute
    
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/firdausmntp/chat-bot-nextjs
    email: support@fsu.my.id
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://chat.fsu.my.id/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
  /chat:
    post:
      tags:
        - Chat
      summary: Send chat message
      description: |
        Send a message to the AI chatbot and receive a streaming response using Server-Sent Events (SSE).
        Supports multi-turn conversations with context and optional image analysis.
      operationId: sendChatMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              simpleChat:
                summary: Simple text chat
                value:
                  messages:
                    - role: user
                      content: "What is artificial intelligence?"
                  model: "resita-chatgpt"
                  temperature: 0.7
              chatWithImage:
                summary: Chat with image context
                value:
                  messages:
                    - role: user
                      content: "What's in this image?"
                  model: "nekolabs-gpt5mini"
                  imageUrl: "/temp/abc123_1234567890.jpg"
                  temperature: 0.8
              multiTurnChat:
                summary: Multi-turn conversation
                value:
                  messages:
                    - role: user
                      content: "Tell me about quantum computing"
                    - role: assistant
                      content: "Quantum computing is a revolutionary..."
                    - role: user
                      content: "Can you explain more about qubits?"
                  model: "resita-claude"
      responses:
        '200':
          description: Successful response with streaming data
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Request limit per window
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests in current window
            X-RateLimit-Reset:
              schema:
                type: integer
              description: Unix timestamp when the rate limit resets
          content:
            text/event-stream:
              schema:
                type: string
                format: binary
              examples:
                streamingResponse:
                  summary: SSE streaming response
                  value: |
                    data: {"type":"start"}
                    
                    data: {"type":"content","text":"Hello!"}
                    
                    data: {"type":"content","text":" How"}
                    
                    data: {"type":"content","text":" can"}
                    
                    data: {"type":"content","text":" I"}
                    
                    data: {"type":"content","text":" help"}
                    
                    data: {"type":"content","text":" you?"}
                    
                    data: {"type":"end"}
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  schemas:
    ChatRequest:
      type: object
      required:
        - messages
      properties:
        messages:
          type: array
          description: Array of conversation messages
          minItems: 1
          items:
            $ref: '#/components/schemas/Message'
        model:
          type: string
          description: AI model to use
          default: "resita-chatgpt"
          enum:
            - resita-aicoding
            - resita-claude
            - resita-chatgpt
            - resita-felo
            - resita-gemini
            - resita-gptlogic
            - resita-venice
            - nekolabs-gpt4o
            - nekolabs-gpt41
            - nekolabs-gpt5mini
            - nekolabs-gpt5nano
        temperature:
          type: number
          description: Controls randomness in generation (0.0 = deterministic, 1.0 = very random)
          minimum: 0.0
          maximum: 1.0
          default: 0.7
        maxTokens:
          type: integer
          description: Maximum number of tokens to generate
          minimum: 1
          maximum: 2048
          default: 1024
        imageUrl:
          type: string
          description: Optional image URL for vision models
          nullable: true
    Message:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          description: Role of the message sender
          enum:
            - user
            - assistant
            - system
        content:
          type: string
          description: Message content
          minLength: 1
        timestamp:
          type: integer
          description: Unix timestamp of the message
          nullable: true
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - code
        - statusCode
      properties:
        error:
          type: string
          description: Error title
        message:
          type: string
          description: Detailed error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_REQUEST
            - FILE_TOO_LARGE
            - INVALID_FILE_TYPE
            - IMAGE_NOT_FOUND
            - RATE_LIMIT_EXCEEDED
            - QUOTA_EXCEEDED
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - INVALID_API_KEY
        statusCode:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        path:
          type: string
          description: API path that caused the error
        requestId:
          type: string
          description: Unique request identifier for tracking
        retryAfter:
          type: integer
          description: Seconds to wait before retrying (for rate limit errors)
  responses:
    BadRequest:
      description: Bad Request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidRequest:
              value:
                error: "Invalid request"
                message: "Messages array is required"
                code: "INVALID_REQUEST"
                statusCode: 400
                timestamp: "2025-10-13T10:30:00.000Z"
                path: "/api/chat"
    RateLimitExceeded:
      description: Too Many Requests - Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            rateLimitExceeded:
              value:
                error: "Rate limit exceeded"
                message: "Too many requests. Please try again in 45 seconds."
                code: "RATE_LIMIT_EXCEEDED"
                statusCode: 429
                retryAfter: 45
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internalError:
              value:
                error: "Internal server error"
                message: "An unexpected error occurred"
                code: "INTERNAL_ERROR"
                statusCode: 500
    ServiceUnavailable:
      description: Service Unavailable - External service is down
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            serviceUnavailable:
              value:
                error: "Service unavailable"
                message: "AI service is temporarily unavailable"
                code: "SERVICE_UNAVAILABLE"
                statusCode: 503
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication (coming soon).
        Include the token in the Authorization header: `Authorization: Bearer YOUR_TOKEN`
security: []
...